Améliore les modules génériques et quelques commentaires de TODO
