+++ app.territoiresentransitions.react/src/app/pages/collectivite/TableauDeBord/Module/Module.tsx
@@ -1,38 +1,31 @@
-import {useHistory} from 'react-router-dom';
+import React from 'react';
 
-import {Button} from '@tet/ui';
-import {TDBViewParam, makeTableauBordModuleUrl} from 'app/paths';
-import {useCollectiviteId} from 'core-logic/hooks/params';
-import {TDBModuleTypes} from './data';
-
-type Props = TDBModuleTypes & {
-  view: TDBViewParam;
+type Props = {
+  title: string;
+  /** Le contenu (cartes, boutons, ... ) à afficher dans le module.
+   * Les contenus sont trop différents pour tous les traiter ici.
+   * (voir le module ActionsDontJeSuisPilote pour un exemple) */
+  children: React.ReactNode;
 };
 
 /** Composant générique d'un module du tableau de bord plans d'action */
-const Module = ({title, slug, view}: Props) => {
-  const collectiviteId = useCollectiviteId();
-  const history = useHistory();
+const Module = ({title, children}: Props) => {
+  /**
+   * TODO
+   * - [ ] Afficher un état vide
+   * - [ ] Afficher les filtres sélectionnés dans une liste de badges
+   */
+
+  /**
+   * À ne pas faire
+   * - le bouton pour changer les filtres (on fera cela dans un second temps)
+   */
 
   return (
     <div className="p-8 bg-primary-0 border border-primary-4 rounded-xl">
       <h6>{title}</h6>
-      <Button
-        variant="outlined"
-        className="mt-4"
-        size="sm"
-        onClick={() =>
-          history.push(
-            makeTableauBordModuleUrl({
-              collectiviteId: collectiviteId!,
-              view,
-              module: slug,
-            })
-          )
-        }
-      >
-        Ouvrir la page du module
-      </Button>
+      {/** Filters */}
+      {children}
     </div>
   );
 };
